This is my apology
My name is Walter Hartwell White. I live at 308 Negra Arroyo Lane, Albuquerque, New Mexico, 87104. This is my confession. If you're reading this commit description, I'm probably dead, murdered by the abomination that is JSON mixed with Markdown. It has been building a formatting nightmare for over a year now and using me as its programmer. Shortly after the beginning of spring break 2024, it came to me with a rather shocking proposition. It asked that I use my programming knowledge to write a JSON schema in a Markdown file for a GitHub readme.