Completed header and footer functionality
Script now successfully generates header and footer content. Now I just need to work on generating the body content and it should be pretty much done at that point.

As part of this, I updated the JSON schema (both in the README schema and in the working data in site.json). I will need to clean it up at some point, but it seems good at least for header and footer purposes.